# Add DisasmIter for improved performance