--- a/bench_iter_test.go
+++ b/bench_iter_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package gapstone
 
 import (