@bentay 2017-06-21
Ah, C.CBytes is apparently a newish feature.