@lunixbochs 2018-08-08
Maybe compare to https://github.com/lunixbochs/capstr performance? I heavily optimized the C->Go interface and allocations, as cgo is a bit slow.